requirements_cuda10.txt
numpy
matplotlib
imageio
pypng
pyyaml
mxnet-cu101==1.5.1
flow_vis
moviepy
opencv-contrib-python   # For cv2. May be also only opencv-python.
tqdm
